# openapi.yaml
openapi: 3.1.0
info:
  title: Predictify API
  version: 0.0.1
  description: Backend API for Predictify — a Stellar/Soroban prediction-markets dApp
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            requestId:
              type: string
          required:
            - code
      required:
        - error
    ValidationErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            details:
              nullable: true
          required:
            - code
      required:
        - error
...
    Market:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        status:
          type: string
        metadata:
          nullable: true
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - question
        - status
        - version
        - createdAt
    MarketSearchResult:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Market'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        page:
          type: integer
        fallback:
          type: boolean
        pagination:
          type: object
          properties:
            limit:
              type: integer
            offset:
              type: integer
            page:
              type: integer
            total:
              type: integer
            fallback:
              type: boolean
          required:
            - limit
            - offset
            - page
            - total
            - fallback
        meta:
          type: object
          properties:
            limit:
              type: integer
            offset:
              type: integer
            page:
              type: integer
            total:
              type: integer
            fallback:
              type: boolean
          required:
            - limit
            - offset
            - page
            - total
            - fallback
      required:
        - data
        - total
        - limit
        - offset
        - page
        - fallback
        - pagination
        - meta
    PatchMarketRequest:
      type: object
      properties:
        question:
          type: string
        metadata:
          nullable: true
        expectedVersion:
          type: integer
          minimum: 0
      required:
        - expectedVersion
...
    FeaturedMarket:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        status:
          type: string
        resolutionOutcome:
          type: string
          nullable: true
        resolutionTime:
          type: string
          format: date-time
        winningOutcome:
          type: string
          nullable: true
        metadata:
          nullable: true
        featuredAt:
          type: string
          nullable: true
          format: date-time
        featuredBy:
          type: string
          nullable: true
      required:
        - id
        - question
        - status
        - resolutionTime
        - featuredAt
        - featuredBy
...
    AuditEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
        actor:
          type: string
        targetAddress:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - action
        - createdAt
...
paths:
...
  /api/markets/search:
    get:
      operationId: searchMarkets
      tags:
        - Markets
      summary: Full-text search across markets
      parameters:
        - schema:
            type: string
            minLength: 1
          required: true
          name: q
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            default: 20
          required: false
          name: limit
          in: query
        - schema:
            type: integer
            nullable: true
            minimum: 0
            default: 0
          required: false
          name: offset
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
          required: false
          name: page
          in: query
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketSearchResult'
        '400':
          description: Missing query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /api/markets/{id}:
    get:
      operationId: getMarketById
      tags:
        - Markets
      summary: Get a market by ID
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        '200':
          description: Market
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Market'
                required:
                  - data
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    patch:
      operationId: updateMarket
      tags:
        - Markets
      summary: Update a market (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchMarketRequest'
      responses:
        '200':
          description: Updated market
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Market'
                required:
                  - data
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Version conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/markets/featured:
    get:
      tags:
        - Markets
      summary: List currently featured markets for the home page
      parameters:
        - schema:
            type: integer
            minimum: 1
            maximum: 20
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Featured markets ordered by most recently featured first
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FeaturedMarket'
                required:
                  - data
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/admin/audit:
    get:
      operationId: getAdminAuditLog
      tags:
        - Admin
      summary: List audit log entries (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
          required: false
          name: action
          in: query
        - schema:
            type: string
          required: false
          name: actor
          in: query
        - schema:
            type: string
            format: date-time
          required: false
          name: startDate
          in: query
        - schema:
            type: string
            format: date-time
          required: false
          name: endDate
          in: query
        - schema:
            type: string
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Paginated audit log
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEntry'
                  nextCursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - nextCursor
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...