# src/routes/markets/index.ts
  
  
/* eslint-disable @typescript-eslint/no-explicit-any */ 
import { Router } from "express";
import { listMarkets, listUpcomingMarkets, getMarketById, updateMarket, VersionConflictError } from "../../services/marketService";
import { searchMarkets } from "../../repositories/marketRepository";
import { requireAdmin, AuthenticatedRequest } from "../../middleware/auth";
import { rateLimitAnon } from "../../middleware/rateLimitAnon";
import { listFeaturedMarkets } from "../../services/marketFeatureService";
import { z } from "zod";
import { logger } from "../../config/logger";
import { disputesRouter } from "../disputes";
import { trendingRouter } from "./trending";
import { marketAuditRouter } from "../marketAudit";

export const marketsRouter = Router();

marketsRouter.use("/:id/disputes", disputesRouter);

marketsRouter.use(rateLimitAnon);
marketsRouter.use("/trending", trendingRouter);

// Per-market audit log: GET /api/markets/:id/audit (#216)
marketsRouter.use("/:id/audit", marketAuditRouter);

const patchMarketSchema = z.object({
  question: z.string().optional(),
  metadata: z.any().optional(),
  expectedVersion: z.number().int().nonnegative(),
}).strict();

marketsRouter.get("/search", async (req, res, next) => {
  const reqId = String((req as any).id ?? "anon");
  try {
    const q = req.query.q as string;
    if (typeof q !== "string" || !q.trim()) {
      logger.warn({ reqId, correlationId: reqId, query: req.query }, "markets_search_validation_failed");
      return res.status(400).json({
        error: {
          code: "validation_error",
          message: "Search query parameter 'q' is required",
          correlationId: reqId,
          requestId: reqId,
        },
      });
    }

    const limit = parseInt(req.query.limit as string) || 20;
    const offset = parseInt(req.query.offset as string) || (parseInt(req.query.page as string) > 1 ? (parseInt(req.query.page as string) - 1) * limit : 0);
    const page = parseInt(req.query.page as string) || Math.floor(offset / limit) + 1;

    logger.info({ reqId, correlationId: reqId, query: q, limit, offset }, "markets_search_executed");

    const result = await searchMarkets({ query: q, limit, offset });

    return res.status(200).json({
      data: result.data,
      total: result.total,
      limit,
      offset,
      page,
      fallback: result.fallback,
      pagination: {
        limit,
        offset,
        page,
        total: result.total,
        fallback: result.fallback,
      },
      meta: {
        limit,
        offset,
        page,
        total: result.total,
        fallback: result.fallback,
      },
    });
  } catch (err) {
    logger.error({ reqId, correlationId: reqId, err }, "markets_search_failed");
    return next(err);
  }
});

/**
 * GET /api/markets - List active markets with pagination
 *
 * Query Parameters:
 *   - limit: number (1-100, default: 20) - max results per page
 *   - offset: number (default: 0) - pagination offset
 *   - page: number (default: 1) - alternative to offset
 *
 * Validation:
 *   - Returns 400 if limit > 100 or is NaN
 *
 * Logging:
 *   - Includes correlation ID from request context
 *   - Logs validation failures and errors with full context
 */
marketsRouter.get("/", async (req, res, next) => {
  const reqId = String((req as any).id ?? "anon");
  try {
    if (req.query.limit !== undefined && (isNaN(Number(req.query.limit)) || Number(req.query.limit) > 100)) {
      logger.warn(
        { reqId, correlationId: reqId, limit: req.query.limit },
        "markets_list_invalid_limit"
      );
      return res.status(400).json({
        error: {
          code: "invalid_query",
          message: "Limit must be a number between 1 and 100",
          correlationId: reqId,
        },
      });
    }

    logger.debug({ reqId, correlationId: reqId, limit: req.query.limit }, "markets_list_fetching");
    const data = await listMarkets();

    logger.info({ reqId, correlationId: reqId, count: data.length }, "markets_list_success");
    return res.json({ data });
  } catch (e) {
    logger.error({ reqId, correlationId: reqId, err: e }, "markets_list_failed");
    return next(e);
  }
});

// Public: curated home-page list. Served ahead of `/:id` so the literal
// path is matched first.
marketsRouter.get("/featured", async (req, res, next) => {
  try {
    const rawLimit = req.query.limit;
    let parsedLimit: number | undefined;
    if (rawLimit !== undefined) {
      const num = Number(rawLimit);
      if (!Number.isFinite(num) || num < 1 || num > 20) {
        return res.status(400).json({
          error: { code: "invalid_query", message: "limit must be an integer between 1 and 20" },
        });
      }
      parsedLimit = Math.floor(num);
    }
    const data = await listFeaturedMarkets(parsedLimit);
    return res.json({ data });
  } catch (e) {
    return next(e);
  }
});

marketsRouter.get("/upcoming", async (req, res, next) => {
  const reqId = String((req as any).id ?? "anon");
  try {
    if (
      req.query.limit !== undefined &&
      (isNaN(Number(req.query.limit)) || Number(req.query.limit) < 1 || Number(req.query.limit) > 100)
    ) {
      return res.status(400).json({
        error: { code: "validation_error", message: "limit must be between 1 and 100", requestId: reqId },
      });
    }
    const limit = req.query.limit !== undefined ? Number(req.query.limit) : 50;
    const data = await listUpcomingMarkets({ limit });
    logger.info({ reqId, correlationId: reqId, count: data.length }, "markets_upcoming_listed");
    return res.json({ data });
  } catch (err) {
    logger.error({ reqId, correlationId: reqId, err }, "markets_upcoming_failed");
    return next(err);
  }
});

marketsRouter.get("/:id", async (req, res, next) => {
  const reqId = String((req as any).id ?? "anon");
  const marketId = req.params.id as string;

  try {
    if (!marketId || typeof marketId !== "string") {
      logger.warn({ reqId, correlationId: reqId, marketId }, "markets_get_invalid_id");
      return res.status(400).json({
        error: {
          code: "invalid_request",
          message: "Market ID is required and must be a string",
          correlationId: reqId,
        },
      });
    }

    logger.debug({ reqId, correlationId: reqId, marketId }, "markets_get_fetching");
    const market = await getMarketById(marketId);

    if (!market) {
      logger.warn({ reqId, correlationId: reqId, marketId }, "markets_get_not_found");
      return res.status(404).json({
        error: {
          code: "not_found",
          message: `Market with ID ${marketId} not found`,
          correlationId: reqId,
        },
      });
    }

    logger.info({ reqId, correlationId: reqId, marketId }, "markets_get_success");
    return res.json({ data: market });
  } catch (e) {
    logger.error({ reqId, correlationId: reqId, marketId, err: e }, "markets_get_failed");
    return next(e);
  }
});

/**
 * PATCH /api/markets/:id - Update a market (admin only)
 *
 * Authorization:
 *   - Requires admin role via requireAdmin middleware
 *
 * Request Body:
 *   - question?: string - new question text
 *   - metadata?: any - custom metadata object
 *   - expectedVersion: number - current version for optimistic locking
 *
 * Responses:
 *   - 200: Updated market object
 *   - 400: Validation error
 *   - 401: Unauthorized
 *   - 404: Market not found
 *   - 409: Version conflict (stale update)
 *   - 500: Database error
 *
 * Optimistic Locking:
 *   - expectedVersion must match current version
 *   - Version incremented on successful update
 *   - 409 response if version mismatch (prevents lost updates)
 *
 * Audit:
 *   - Change logged in marketAuditLog table
 *   - Includes before/after state and admin address
 *
 * Logging:
 *   - Includes correlation ID, market ID, admin address, and version info
 */
marketsRouter.patch("/:id", requireAdmin, async (req: AuthenticatedRequest, res, next) => {
  const reqId = String((req as any).id ?? "anon");
  const marketId = req.params.id as string;
  const adminAddress = req.user?.stellarAddress;

  try {
    // Validate schema
    const parsed = patchMarketSchema.safeParse(req.body);
    if (!parsed.success) {
      logger.warn(
        {
          reqId,
          correlationId: reqId,
          marketId,
          adminAddress,
          issues: parsed.error.issues,
        },
        "markets_patch_validation_failed"
      );
      return res.status(400).json({
        error: {
          code: "validation_error",
          message: "Invalid request body",
          details: parsed.error.issues,
          correlationId: reqId,
        },
      });
    }

    const { question, metadata, expectedVersion } = parsed.data;

    // Build patch object
    const patch: { question?: string; metadata?: any } = {};
    if (question !== undefined) patch.question = question;
    if (metadata !== undefined) patch.metadata = metadata;

    logger.info(
      {
        reqId,
        correlationId: reqId,
        marketId,
        adminAddress,
        expectedVersion,
        fieldsUpdated: Object.keys(patch),
      },
      "markets_patch_updating"
    );

    const updated = await updateMarket(marketId, patch, expectedVersion, adminAddress!);

    logger.info(
      {
        reqId,
        correlationId: reqId,
        marketId,
        adminAddress,
        newVersion: updated.version,
      },
      "markets_patch_success"
    );
    return res.json({ data: updated });
  } catch (e) {
    if (e instanceof VersionConflictError) {
      logger.warn(
        {
          reqId,
          correlationId: reqId,
          marketId,
          adminAddress,
        },
        "markets_patch_version_conflict"
      );
      return res.status(409).json({
        error: {
          code: "version_conflict",
          message: "Market has been modified by another request. Please refresh and try again.",
          correlationId: reqId,
        },
      });
    }

    if ((e as any).status === 404) {
      logger.warn({ reqId, correlationId: reqId, marketId, adminAddress }, "markets_patch_not_found");
      return res.status(404).json({
        error: {
          code: "not_found",
          message: `Market with ID ${marketId} not found`,
          correlationId: reqId,
        },
      });
    }

    logger.error(
      { reqId, correlationId: reqId, marketId, adminAddress, err: e },
      "markets_patch_failed"
    );
    return next(e);
  }
});